Fix FastGA command logging to show actual arguments
Previously, the log output showed a hardcoded format that didn't include
conditional parameters like -l (min_length) and -i (min_identity) when they
were non-zero. This was misleading as the parameters were being passed to
FastGA but not shown in the logs.

Changes:
- Build the command first with all conditional arguments
- Collect actual args and log them using cmd.get_args()
- Now logs show the exact command being executed including -l, -i, etc.

This applies to both align_to_1aln() and run_alignment() methods.